+++ src/agnocastlib/include/agnocast/bridge/standard/agnocast_standard_bridge_manager.hpp
@@ -69,14 +69,12 @@ private:
   void send_delegation(const MqMsgBridge & req, pid_t owner_pid);
   void process_managed_bridge(
     const std::string & topic_name, const std::optional<MqMsgBridge> & req);
+  std::pair<bool, bool> should_remove_bridge(const std::string & topic_name, bool is_r2a);
 
   void check_parent_alive();
   void check_active_bridges();
   void check_managed_bridges();
   void check_should_exit();
-
-  void remove_active_bridge(
-    const std::string & topic_name_with_direction, bool keep_managed = false);
 
   static std::pair<std::string, std::string> extract_topic_info(const MqMsgBridge & req);
 };